task 24 minor syntax improvement

+++ supervised_learning/classification/24-one_hot_encode.py
@@ -15,7 +15,7 @@ def one_hot_encode(Y, classes):
     import numpy as np
 
     biggestPossibleClassNumber = max(Y)
-    parent = np.empty(biggestPossibleClassNumber, classes)
+    parent = np.empty(biggestPossibleClassNumber.astype(int), classes)
 
     for numberCheckedFor in range(0, biggestPossibleClassNumber):
         parent[numberCheckedFor] = \
